Find valid parsers for each header
This commit adds functionality to go over each field in each row, and
process parse it with every known parser to find out which ones work.

Whenever a parse fails for a field for a given header, we remove
it from the valid pool of parsers for said header. This way, we avoid
re-running already invalid parsers over the rest of the dataset for that
column.

It would be possible to also keep track of what field caused each parser
to fail, for debugging purposes. Such feature might be added later on
behind an CLI argument.