Fix handling of delimiter argument.
Before, the type of the `delimiter` argument was `u8`, which is the type
that `csv::ReaderBuilder.delimiter` expects. But that makes it harder
to work with `clap`. For example, the default value would be attempted
to be parsed as `44` rather than `','`, panicking.

With this change, we instead parse and operate the argument as `char`,
and convert to `u8` at the last moment.