Use `quote`, `syn`, and `prettyplease` to generate code.
Rather than using a janky string concatenation method (that may or may
not actually be clearer), this commit changes the code generation to
use the `quote` and `syn` crates. This makes the code generation logic
to resemble real code a bit more, and should be able to better catch
issues.

There is one disadvantage so far: The ability to intersperse empty lines
between the fields of the output struct has been lost, which might make
it less clean. But maybe there's a way to recover that if I used a
crate different than `prettyplease` to do the formatting.