Add `min_fields` argument
This change adds a way to skip records that have less than a specified
amount of fields. This can be useful when parsing files that might have
some records used as section headers. e.g. with a single field just to
separate from other sections.

I considered two strategies for this:
- Check the amount of non-empty fields in each record.
- Trim empty fields from the end, and then see how many are left.

If this program has a special need to be more exhaustive in terms of
parsing the rows, the second one might be more useful. But mainly the
resulting structs just need to be approximately good. It is easy to
touch a field or two in the code to adjust if things fail to parse. But
I still want to avoid having all fields marked as optional because of a
header, so I think the first strategy works well enough.