Add option --blank-lines to space out the fields
When generating a larger struct, and, in particular, when fields
might have the `#[serde(rename = ...)]` attribute, things can get quite
difficult to parse. I used to manually add empty spaces, but it is
getting a bit repetitive. So now `csv2serde` can do it for me!

I looked around for a way to do this with `prettyplease`, or even if
I switched over to use `rustfmt`. Unfortunately, I found nothing.
So instead, there's some custom logic that splits the generated code and
intersperses some newline characters wherever it seemed reasonable.