Move default_value for --lines into the attribute
The code was already defaulting `lines` to `usize::MAX`, but it looked
weird in the help message to specify this value. Thankfully, `clap`
allows hiding this value via the `hide_default_value` attribute.

To offer some guidance in the help, the argument message has been
edited to mention: "[default: ALL]".